Move compatibility verifier to a separate module
