+++ pinot-compatibility-verifier/src/main/java/org/apache/pinot/compat/CompatTestOperation.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.pinot.compat.tests;
+package org.apache.pinot.compat;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import java.util.List;
