+++ pom.xml
@@ -55,6 +55,7 @@
     <module>pinot-distribution</module>
     <module>pinot-connectors</module>
     <module>pinot-segment-local</module>
+    <module>pinot-compatibility-verifier</module>
   </modules>
 
   <licenses>
